Pass line hash instead of params
Changes:

* Changes from name, params, value params to using a hash
* Also tidies up tests a bit with a helper method for parsing a line with
a builder.

Notes:

I'm always hesitant to make changes like this which extract duplicate code into a test helper, but the internals of the helper method are not really relevant to the test. All that the user needs to know is that the helper sends the line parameters appropriately to the builder, and then it can test to make sure things got set right on the other end.